+++ example/lib/pages/select_bloc.dart
@@ -19,7 +19,7 @@ class _PageSelectBlocState extends State<PageSelectBloc> {
       print("Change type: ${change.type}");
       print("Number of items impacted: ${change.value}");
       print("Query: ${change.query}");
-      if (change.type == DatabaseChangeType.update) {
+      if (change.type == DatabaseChange.update) {
         print("${change.value} items updated");
       }
     });